1. Implement: PATCH /api/shop/product 2. Remove: /fs/
