--- src/route/fs/mod.rs
@@ -1,16 +1,0 @@
-use warp::{
-    Filter,
-    reply::Reply,
-    filters::BoxedFilter,
-    path,
-};
-use crate::state::State;
-
-mod shop;
-
-pub fn filter(state: BoxedFilter<(State,)>) -> BoxedFilter<(impl Reply,)> {
-    path("shop").and(
-        shop::filter(state.clone())
-    )
-    .boxed()
-}
