--- src/route/fs/shop/product/mod.rs
@@ -1,18 +1,0 @@
-use warp::{
-    Filter,
-    reply::Reply,
-    filters::BoxedFilter,
-    path,
-};
-use crate::{
-    state::State,
-};
-
-mod image;
-
-pub fn filter(state: BoxedFilter<(State,)>) -> BoxedFilter<(impl Reply,)> {
-    path("image").and(
-        image::filter(state.clone())
-    )
-    .boxed()
-}
